+++ infra/in_memory/products.py
@@ -28,7 +28,7 @@ class ProductsInMemory:
     def read_all(self) -> list[Product]:
         return list(self.products.values())
 
-    def update_price(self, product_id: UUID, new_price: int) -> None:
+    def update_price(self, product_id: UUID, new_price: float) -> None:
         try:
             self.products[product_id]
         except KeyError:
